feat(chat): 新增聊天视图相关功能与组件
- 添加 MemorySchemaEditor 和 MemorySchemaFieldEditor 组件
- 实现聊天会话生命周期管理钩子
- 新增聊天视图展示逻辑和UI控制
- 添加聊天消息处理和流式传输功能
- 实现聊天机器人管理和表单交互
- 添加聊天初始化和会话管理功能
- 实现聊天消息管道和状态管理
